requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg[binary]
aiofiles
email-validator
